:sparkles: feat: add meal controller

+++ src/controllers/meal.controller.ts
@@ -1,17 +1,26 @@
+import fs from "fs";
+import path from "path";
 import { PrismaClient } from "@prisma/client";
 import { Request, Response } from "express";
+import ImageUpload from "../services/image-uploader";
 
 const prisma = new PrismaClient();
+const imageUpload = new ImageUpload();
 
 const MealController = {
     async createMeal(req: Request, res: Response) {
         try {
-            const { userId, description, calories, datetime } = req.body;
+            const { userId, title, description, calories } = req.body;
+            let imagePath = null;
 
-            if (!userId || !datetime) {
-                return res.status(400).json({
-                    error: "ID de usuário e data/hora da refeição são obrigatórios",
-                });
+            if (req.file) {
+                imagePath = `../../uploads/meals/${req.file.filename}`;
+            }
+
+            if (!userId) {
+                return res
+                    .status(400)
+                    .json({ error: "ID do usuário é obrigatório." });
             }
 
             const newMeal = await prisma.meal.create({
@@ -19,9 +28,11 @@ const MealController = {
                     user: {
                         connect: { id: userId },
                     },
+                    title,
                     description,
-                    calories,
-                    datetime: new Date(datetime),
+                    calories: parseInt(calories, 10),
+                    datetime: new Date(),
+                    image: imagePath,
                 },
             });
 
@@ -29,6 +40,53 @@ const MealController = {
         } catch (error) {
             console.error("Erro ao criar refeição:", error);
             return res.status(500).json({ error: "Erro ao criar refeição" });
+        }
+    },
+
+    async updateMeal(req: Request, res: Response) {
+        try {
+            const mealId = req.params.mealId;
+            const { title, description, calories } = req.body;
+            let imagePath = null;
+
+            if (req.file) {
+                imagePath = `../../uploads/meals/${req.file.filename}`;
+            }
+
+            const meal = await prisma.meal.findUnique({
+                where: { id: parseInt(mealId) },
+            });
+
+            if (!meal) {
+                return res
+                    .status(404)
+                    .json({ error: "Refeição não encontrada" });
+            }
+
+            if (imagePath && meal.image) {
+                const oldImagePath = path.join(__dirname, meal.image);
+                if (fs.existsSync(oldImagePath)) {
+                    fs.unlinkSync(oldImagePath);
+                }
+            }
+
+            const updatedMeal = await prisma.meal.update({
+                where: { id: parseInt(mealId) },
+                data: {
+                    title,
+                    description,
+                    calories: parseInt(calories, 10),
+
+                    image: imagePath || meal.image,
+                },
+            });
+
+            return res.status(200).json(updatedMeal);
+        } catch (error) {
+            console.error("Erro ao atualizar refeição:", error);
+            return res
+                .status(500)
+                .json({ error: "Erro ao atualizar refeição" });
         }
     },
 
@@ -64,32 +122,26 @@ const MealController = {
         }
     },
 
-    async updateMeal(req: Request, res: Response) {
-        try {
-            const mealId = req.params.mealId;
-            const { description, calories, datetime } = req.body;
-
-            const updatedMeal = await prisma.meal.update({
-                where: { id: parseInt(mealId) },
-                data: {
-                    description,
-                    calories,
-                    datetime: datetime ? new Date(datetime) : undefined,
-                },
-            });
-
-            return res.status(200).json(updatedMeal);
-        } catch (error) {
-            console.error("Erro ao atualizar refeição:", error);
-            return res
-                .status(500)
-                .json({ error: "Erro ao atualizar refeição" });
-        }
-    },
-
     async deleteMeal(req: Request, res: Response) {
         try {
             const mealId = req.params.mealId;
+
+            const meal = await prisma.meal.findUnique({
+                where: { id: parseInt(mealId) },
+            });
+
+            if (!meal) {
+                return res
+                    .status(404)
+                    .json({ error: "Refeição não encontrada" });
+            }
+
+            if (meal.image) {
+                const imagePath = path.join(__dirname, meal.image);
+                if (fs.existsSync(imagePath)) {
+                    fs.unlinkSync(imagePath);
+                }
+            }
 
             await prisma.meal.delete({
                 where: { id: parseInt(mealId) },
